BE/feat/redirectUrl수정

See merge request s09-ai-image-sub2/S09P22A606!101

+++ project/server/auth/src/main/java/com/cc/auth/domain/auth/dto/LogoutResponseDto.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @Builder
 public class LogoutResponseDto {
 
-    private Long memberid;
+    private int memberid;
 
     public static LogoutResponseDto of(Member member){
         return LogoutResponseDto.builder()
